LayoutTests/imported/w3c/web-platform-tests/websockets/Send-paired-surrogates.any-expected.txt


FAIL Send paired surrogates data on a WebSocket - Connection should be closed assert_equals: expected 0 but got 4
